Removed the duplication of certain methods in the test classes. Now new seekers should become easier to test.
