+++ test/models/seeker/LoudnessSeekerTest.java
@@ -3,8 +3,6 @@ package models.seeker;
 import java.util.ArrayList;
 import java.util.List;
 
-import basic.BasicTest;
-import models.record.Track;
 import models.score.ScoreStorage;
 import models.score.ScoreMap;
 
@@ -24,17 +22,12 @@ import static org.junit.Assert.assertEquals;
  * @author stefan boodt
  *
  */
-public class LoudnessSeekerTest extends BasicTest {
+public class LoudnessSeekerTest extends AbstractSeekerTest {
 
     /**
      * The seeker under test.
      */
     private LoudnessSeeker seeker;
-    
-    /**
-     * Default track used by the tests.
-     */
-    protected static final Track DEFAULT_TRACK = new Track(10, 1000);
     
     @Override
     public void setUp() throws Exception {
@@ -55,7 +48,7 @@ public class LoudnessSeekerTest extends BasicTest {
      * @param seeker the new seeker.
      */
     public void setSeeker(final LoudnessSeeker seeker) {
-        setObjectUnderTest(seeker);
+        super.setSeeker(seeker);
         this.seeker = seeker;
     }
 
@@ -72,24 +65,6 @@ public class LoudnessSeekerTest extends BasicTest {
             doubles.add(f);
         }
         return doubles;
-    }
-
-    /**
-     * Tests the {@link LoudnessSeeker#getTrack()} method.
-     */
-    @Test
-    public void testGetTrack() {
-        assertEquals(DEFAULT_TRACK, getSeeker().getTrack());
-    }
-
-    /**
-     * Tests the {@link LoudnessSeeker#getTrack()} method.
-     */
-    @Test
-    public void testGetTrackAfterSetting() {
-        final Track expected = new Track(11, -1);
-        getSeeker().setTrack(expected);
-        assertEquals(expected, getSeeker().getTrack());
     }
     
     /**
